[eclipse/xtext-eclipse#299] Added support for Language Server Jar
The IDE project's build can be extended to create an executable Language
Server jar. For Maven based builds the maven-shade-plugin is used and
for Gradle based builds the ShadowJar plugin is used.
Due to a limitation of the maven-shade-plugin in a Maven/Tycho
environment the build must be tweaked a bit to materialize dependencies
with the maven-dependencies-plugin and attached for inclusion with the
addjars-maven-plugin.

Extended the test suite for the different configurations with language
server support:
- plain Maven
- Maven/Tycho
- Gradle

Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>